move to parse-utils.
